style(LottoTests) : 폴더이동
