--- src/test/kotlin/lotto/LottoTest.kt
@@ -1,35 +1,0 @@
-package lotto
-
-import lotto.domain.model.Lotto
-import lotto.domain.model.WinningLottery
-import org.assertj.core.api.Assertions.assertThat
-import org.junit.jupiter.api.DisplayName
-import org.junit.jupiter.api.Nested
-import org.junit.jupiter.api.Test
-import org.junit.jupiter.api.assertThrows
-
-
-class LottoTest {
-
-    @Test
-    fun `로또 번호의 개수가 6개가 넘어가면 예외가 발생한다`() {
-        assertThrows<IllegalArgumentException> {
-            Lotto(listOf(1, 2, 3, 4, 5, 6, 7))
-        }
-    }
-
-    @Test
-    fun `로또 번호에 중복된 숫자가 있으면 예외가 발생한다`() {
-        assertThrows<IllegalArgumentException> {
-            Lotto(listOf(1, 2, 3, 4, 5, 5))
-        }
-    }
-
-    @Test
-    fun `로또 번호가 1~45사이 숫자가 아니면 예외가 발생한다`() {
-        assertThrows<IllegalArgumentException> {
-            Lotto(listOf(0, 2, 3, 4, 5, 48))
-        }
-    }
-
-}
